+++ libs/libc/stdlib/lib_arc4random.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * net/utils/net_getrandom.c
+ * libs/libc/stdlib/lib_arc4random.c
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -24,9 +24,9 @@
  * Included Files
  ****************************************************************************/
 
-#include <nuttx/config.h>
-
+#include <errno.h>
 #include <stdint.h>
+#include <stdlib.h>
 #include <string.h>
 #include <sys/param.h>
 #include <sys/random.h>
@@ -35,33 +35,68 @@
 #include <nuttx/hashtable.h>
 
 /****************************************************************************
+ * Private Functions
+ ****************************************************************************/
+
+#if defined(CONFIG_DEV_URANDOM) || defined(CONFIG_DEV_RANDOM)
+static int getrandom_all(FAR void *buf, size_t size, int flags)
+{
+  FAR char *tmp = buf;
+
+  while (size > 0)
+    {
+      ssize_t ret = getrandom(tmp, size, flags);
+      if (ret < 0)
+        {
+          if (get_errno() == EINTR)
+            {
+              continue;
+            }
+
+          return ret;
+        }
+
+      tmp += ret;
+      size -= ret;
+    }
+
+  return 0;
+}
+#endif
+
+/****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: net_getrandom
+ * Name: arc4random_buf
  *
  * Description:
- *   Fill a buffer of arbitrary length with randomness. This function is
- *   guaranteed to be success.
+ *   Fill a buffer of arbitrary length with randomness. This is the
+ *   preferred interface for getting random numbers. The traditional
+ *   /dev/random approach is susceptible for things like the attacker
+ *   exhausting file descriptors on purpose.
+ *
+ *   Note that this function cannot fail, other than by asserting.
  *
  * Input Parameters:
  *   bytes  - Buffer for returned random bytes
  *   nbytes - Number of bytes requested.
  *
+ * Returned Value:
+ *   None
+ *
  ****************************************************************************/
 
-void net_getrandom(FAR void *bytes, size_t nbytes)
+void arc4random_buf(FAR void *bytes, size_t nbytes)
 {
 #if defined(CONFIG_DEV_URANDOM) || defined(CONFIG_DEV_RANDOM)
-  ssize_t ret = getrandom(bytes, nbytes, 0);
-
-  if (ret < 0)
+  if (getrandom_all(bytes, nbytes, GRND_RANDOM) >= 0)
     {
-      ret = getrandom(bytes, nbytes, GRND_RANDOM);
+      return;
     }
 
-  if (ret == nbytes)
+  if (getrandom_all(bytes, nbytes, 0) >= 0)
     {
       return;
     }
@@ -81,4 +116,28 @@ void net_getrandom(FAR void *bytes, size_t nbytes)
       nbytes -= ncopy;
       bytes   = (FAR uint8_t *)bytes + ncopy;
     }
+}
+
+/****************************************************************************
+ * Name: arc4random
+ *
+ * Description:
+ *   Returns a single 32-bit value. This is the preferred interface for
+ *   getting random numbers. The traditional /dev/random approach is
+ *   susceptible for things like the attacker exhausting file
+ *   descriptors on purpose.
+ *
+ *   Note that this function cannot fail, other than by asserting.
+ *
+ * Returned Value:
+ *   a random 32-bit value.
+ *
+ ****************************************************************************/
+
+uint32_t arc4random(void)
+{
+  uint32_t ret;
+
+  arc4random_buf(&ret, sizeof(ret));
+  return ret;
 }
