+++ libs/libc/uuid/lib_uuid_create.c
@@ -24,38 +24,9 @@
  * Included Files
  ****************************************************************************/
 
-#include <sys/random.h>
 #include <errno.h>
 #include <stdlib.h>
 #include <uuid.h>
-
-/****************************************************************************
- * Private Functions
- ****************************************************************************/
-
-static int uuid_getrandom(FAR void *buf, size_t size, int flags)
-{
-  FAR char *tmp = buf;
-
-  while (size > 0)
-    {
-      ssize_t ret = getrandom(tmp, size, flags);
-      if (ret < 0)
-        {
-          if (get_errno() == EINTR)
-            {
-              continue;
-            }
-
-          return ret;
-        }
-
-      tmp += ret;
-      size -= ret;
-    }
-
-  return 0;
-}
 
 /****************************************************************************
  * Public Functions
@@ -74,24 +45,7 @@ static int uuid_getrandom(FAR void *buf, size_t size, int flags)
 
 void uuid_create(FAR uuid_t *u, FAR uint32_t *status)
 {
-  int ret;
-
-  ret = uuid_getrandom(u, sizeof(uuid_t), GRND_RANDOM);
-  if (ret < 0)
-    {
-      ret = uuid_getrandom(u, sizeof(uuid_t), 0);
-    }
-
-  if (ret < 0)
-    {
-      FAR unsigned long *beg = (FAR unsigned long *)u;
-      FAR unsigned long *end = (FAR unsigned long *)(u + 1);
-
-      while (beg < end)
-        {
-          *beg++ = rand();
-        }
-    }
+  arc4random_buf(u, sizeof(uuid_t));
 
   u->clock_seq_hi_and_reserved &= ~(1 << 6);
   u->clock_seq_hi_and_reserved |= (1 << 7);
